split ty.rs into smaller parts
